Combine pause menu and settings
