+++ Assets/_BForBoss/Scripts/Managers/WorldManager.cs
@@ -1,6 +1,5 @@
 using Sirenix.OdinInspector;
 using UnityEngine;
-using UnityEngine.InputSystem;
 using UnityEngine.Rendering;
 
 namespace BForBoss
@@ -15,8 +14,7 @@ namespace BForBoss
         [Title("User Interface")]
         [SerializeField] private PauseMenu _pauseMenu;
         [SerializeField] private TimerViewBehaviour _timerView = null;
-        [SerializeField] private SettingsViewBehaviour _settingsViewBehaviour = null;
-        [SerializeField] private ForcedInputUsernameViewBehaviour _forcedUploadView = null;
+        [SerializeField] private ForcedSetUsernameViewBehaviour _forcedUploadView = null;
         [SerializeField] private LeaderboardPanelBehaviour _leaderboardPanel = null;
 
         [Title("Effects")] 
@@ -27,7 +25,6 @@ namespace BForBoss
         [SerializeField] private GameObject _debugCanvas;
 #endif
 
-        private const Key PauseKey = Key.Escape;
         // This probably best placed inside its own utility section
         private readonly StateManager _stateManager = StateManager.Instance;
         private readonly PerigonAnalytics _perigonAnalytics = PerigonAnalytics.Instance;
@@ -62,13 +59,10 @@ namespace BForBoss
 
             if (FindObjectOfType<DebugWindow>() == null)
             {
-                DebugWindow debugWindow = Instantiate(_debugCanvas).gameObject.GetComponent<DebugWindow>();
+                Instantiate(_debugCanvas).gameObject.GetComponent<DebugWindow>();
             }
 #endif
-            
-            
             _character = _player;
-            _pauseMenu.Initialize(_player);
             _postProcessingVolumeWeightTool = new PostProcessingVolumeWeightTool(_deathVolume, 0.1f, 0f, 0.1f);
             _uploadPlayerScoreDataSource = new UploadPlayerScoreDataSource();
         }
@@ -76,35 +70,24 @@ namespace BForBoss
         private void Start()
         {
             _perigonAnalytics.StartSession(SystemInfo.deviceUniqueIdentifier);
-            _player.Initialize();
-            _checkpointManager.Initialize(_detectInput, _timeManagerViewModel);
-            _timeManager.Initialize(_timeManagerViewModel);
-            _timerView.Initialize(_timeManagerViewModel);
-            _settingsViewBehaviour.Initialize(_player);
-            SetupLeaderboardViews();
-            
+            SetupSubManagers();
+            SetupUserInterface();
             _stateManager.SetState(State.PreGame);
         }
 
-        private void SetupLeaderboardViews()
+        private void SetupSubManagers()
         {
-            _leaderboardPanel.Initialize(LockMouseUtility.Instance);
-            _forcedUploadView.Initialize();
+            _player.Initialize();
+            _checkpointManager.Initialize(_detectInput, _timeManagerViewModel);
+            _timeManager.Initialize(_timeManagerViewModel);
         }
 
-        private void Update()
+        private void SetupUserInterface()
         {
-            if (Keyboard.current[PauseKey].wasPressedThisFrame)
-            {
-                if (_pauseMenu.gameObject.activeSelf)
-                {
-                    _pauseMenu.ClosePanel();
-                }
-                else
-                {
-                    _pauseMenu.OpenPanel();
-                }
-            }
+            _pauseMenu.Initialize(_player);
+            _timerView.Initialize(_timeManagerViewModel);
+            _leaderboardPanel.Initialize(LockMouseUtility.Instance);
+            _forcedUploadView.Initialize();
         }
 
         private void OnApplicationQuit()
