--- Assets/_BForBoss/_Leaderboard/Scripts/Username/InputUsernameViewBehaviour.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 63f00337aa94e7744b876aabf78e95c6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
