+++ Assets/_BForBoss/_Leaderboard/UsernameInputPanel.prefab
@@ -121,7 +121,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   _usernameField: {fileID: 5405225908374458249}
   _infoSettingsLabel: {fileID: 5405225907579402478}
-  _saveUsernameButton: {fileID: 5405225907177277454}
+  _setUsernameButton: {fileID: 0}
 --- !u!1 &5405225907094841019
 GameObject:
   m_ObjectHideFlags: 0
@@ -269,7 +269,7 @@ GameObject:
   - component: {fileID: 5405225907177277449}
   - component: {fileID: 5405225907177277454}
   m_Layer: 5
-  m_Name: UploadScoreButton
+  m_Name: SetUsernameButton
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
